Rename the test file
Put test_ is at beginning and remove _tests from the end.  That way when I run coverage for the entire tests directory, it runs the tests in that file.
